# brontide: eliminate all allocations from WriteMessage+Flush