@Roasbeef 2025-09-02
> So the net effect is:
> We save 4 microseconds per message sent.


This is more about entirely eliminating allocations when we write messages, the minor speed up isn't the main target at all. 

> We preallocate 64 KB for every peer.

Correct that we pre-allocate 64 KB per peer, but this memory allocation is now bounded, and doesn't need to later be marked by the GC to be freed (cycles saved traversing the object graph). Memory utilization will now be smoothed out, dampening the peaks you'd see in a typical graph of allocated memory over time for GC'd languages. A node with 1k active connection does indeed see _static_ increase of 64 MB, but that doesn't tell the whole story. 

Before we'd allocate each time (5x per write message) each time we went to write out a message to the peer (individual buffers for: nonce, header, body, and packet size -- so more objects for the GC to chase). 

Rather than focusing on that single snapshot of allocation, we should consider the rate of allocation over time. If a single peer sends us a 65 KB message, each time we're generating garbage that later needs to be cleaned up, which adds overhead for the GC. The overhead we save isn't necessarily showing up in this profile, as it's a short lived benchmark. On the macro level, we increase the overall performance of the entire daemon, as less time is spent on GC cycles (stop-the-world).