@morehouse 2025-09-04
> You're not wrong, but the stated goal is zero (heap) allocs, so then you will have to pre-allocate the max message size.

Ultimately the goal is improved performance.  Eliminating heap allocs is one way to affect that goal, but this is likely a case where there's no performance benefit 99.9% of the time since GC is already optimized efficient for the small, infrequent, short-lived allocations that are being eliminated.

There may actually be a performance *regression* when not under DoS attack, since we're now creating large long-lived allocations for every peer instead of small short-lived ones.  Every time GC occurs, those large allocations need to be scanned in their entirety, causing GC pauses to last longer.

And regardless of the performance improvement or regression, there's a guaranteed cost in memory overhead that this change introduces.  This overhead needs to be weighed against whatever potential performance benefit there may be.

This is why we need good performance data to drive the optimization cycle.  Guessing often produces incorrect results, while increasing code complexity.