@morehouse 2025-09-15
@Roasbeef:  Thanks for collecting more data.  I don't think the data addresses any of my expressed concerns though.  Key information needed to properly optimize:

- What is the "stress test program"?
- How do the variants handle regular traffic (no stress test)?
- A **control group** (i.e. master branch) is needed.
- **CPU profiles** of each variant are critical to determine if the CPU-memory tradeoff is good.
- Ideally provide the SVGs for the web visualizations of all CPU and heap profiles, so we can study them ourselves.

I expressed concerns that this PR increases memory usage vs master.  So we need **memory usage data for master** to properly evaluate that.

I also expressed concerns that this PR will perform poorly under regular traffic, potentially even increasing CPU usage.  We need **CPU profiles** to evaluate that.

We especially need CPU data for **time spent during GC**.  For example, if the master branch only spends 5% of its CPU time in GC, then eliminating allocations across *all of LND* could only increase performance by ~5%!  If that's the case, increasing code complexity to eliminate a small fraction of allocations was silly from the start, and all the discussion on this PR has wasted everyone's time.

-------------------

@yyforyongyu:  I am skeptical that using the buffer pool will provide good memory or CPU benefit in practice.  Why create our own garbage collection scheme instead of relying on Go's optimized garbage collector?  Implementing a GC is tricky, which @Roasbeef's "PR with buf pools" clearly demonstrates -- it *increases* memory usage due to slow memory release.

The "PR with fixed buf pools" makes a memory-time tradeoff --- it imposes a strong memory limit by introducing blocking behavior.  We really need data about how much additional latency is added by this change to decide if the tradeoff is worthwhile.  Since previously there was a maximum 1ms pause and now it blocks forever, this could have a major impact on node responsiveness to peer messages.

> However, the 24-hour timeout seems too aggressive and can be replaced with a much short value like seconds.

Even a few seconds of blocking would cause a huge increase in send latency, something like 100x.  If this happens frequently, it could be a big problem for users.

------------------

There's another variant that I think would be worth benchmarking.  It would be the simplest variant considered so far, and I think it will provide the same or better memory usage in practice as any of the other variants.

[0001-brontide-free-pending-message-after-flushing.patch](https://github.com/user-attachments/files/22346280/0001-brontide-free-pending-message-after-flushing.patch)

The main idea is to work *with* the garbage collector rather than trying to avoid it.  We just drop the reference to the pending message as soon as we no longer need it, so that we no longer prevent the garbage collector from reclaiming that buffer.  This is a big improvement over master, which currently keeps a reference until a new message needs to be written out or until the entire Brontide is garbage collected.

This should provide memory benefit both under DoS and with regular usage.  No new CPU overhead.